requirements.txt
moderngl>=5.8.0
numpy>=1.24.0
Pillow>=10.0.0
ffmpeg-python>=0.2.0
PyYAML>=6.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
